Bugfix...
git-svn-id: https://mem-debug.googlecode.com/svn/trunk@6 8e54e14e-2de8-11df-bba7-995ec2794b1c

+++ source/lib/libmemdebug.c
@@ -738,7 +738,7 @@ static void memdebug_print_object( struct memdebug_object * object )
         "# - Allocated in file:       %s\n"
         "# - Allocated at line:       %i\n"
         "# \n",
-        ( unsigned long int )object->ptr,
+        object->ptr,
         ( unsigned long int )object->size,
         object->alloc_func,
         #ifdef __GNUC__
